[test] Improve typescript-to-proptypes test suite
